requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
Werkzeug==3.1.3
prometheus-client==0.19.0
redis==5.0.1
psycopg2-binary==2.9.9
boto3==1.34.0
gunicorn==21.2.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
safety==2.3.5
